refactor: 컴포넌트, 타입, 도메인, 서비스, 컨트롤러 등 폴더 구조 변경
