+++ src/app/component/RestaurantAddForm/RestaurantAddForm.ts
@@ -8,11 +8,11 @@ import {
   createOptionItems,
   createSelectBox,
   createTextArea,
-} from '../../../util/createFormElement';
-import { RestaurantDataType } from '../../../type/restaurantTypes';
-import { Category, DistanceByWalk } from '../../../enum/enums';
-import handleError from '../../../util/handleError';
-import RestaurantValidator from '../../../validator/RestaurantValidator';
+} from '../../util/createFormElement';
+import { RestaurantDataType } from '../../type/restaurantTypes';
+import { Category, DistanceByWalk } from '../../enum/enums';
+import handleError from '../../util/handleError';
+import RestaurantValidator from '../../validator/RestaurantValidator';
 
 type RestaurantAddFormType = {
   parentComponent: object;
