+++ src/app/component/RestaurantItem/RestaurantItem.css
@@ -1,4 +1,4 @@
-.restaurant__favorite-button {
+/* .restaurant__favorite-button {
   position: absolute;
   top: 16px;
   right: 8px;
@@ -12,4 +12,4 @@
 
 .restaurant__favorite-button.favorited {
   background-image: url('../../../asset/image/favorite-icon-filled.png');
-}
+} */
