+++ src/app/component/RestaurantListFilter/RestaurantListFilter.ts
@@ -1,6 +1,6 @@
-import { Category, SortOrder } from '../../../enum/enums';
-import { createOptionItems } from '../../../util/createFormElement';
-import { $ } from '../../../util/domSelector';
+import { Category, SortOrder } from '../../enum/enums';
+import { createOptionItems } from '../../util/createFormElement';
+import { $ } from '../../util/domSelector';
 
 export default class RestaurantListFilter extends HTMLElement {
   connectedCallback() {
