+++ src/app/domain/Restaurant.ts
