+++ src/app/type/compareFunctionType.ts
